refactored all vessim.core.simulator classes to vessim.TimeSeriesApi
