+++ examples/cosim_example.py
@@ -7,9 +7,9 @@ import argparse
 import mosaik  # type: ignore
 
 from _data import load_carbon_data, load_solar_data
+from vessim import TimeSeriesApi
 from vessim.core.consumer import ComputingSystem, MockPowerMeter
 from vessim.core.microgrid import SimpleMicrogrid
-from vessim.core.simulator import Generator, CarbonApi
 from vessim.core.storage import SimpleBattery, DefaultStoragePolicy
 
 COSIM_CONFIG = {
@@ -57,12 +57,13 @@ def run_simulation(carbon_aware: bool, result_csv: str):
     # Initialize solar generator
     solar_sim = world.start("Generator", sim_start=SIM_START)
     solar = solar_sim.Generator(
-        generator=Generator(actual=load_solar_data(sqm=0.4 * 0.5))
+        generator=TimeSeriesApi(actual=load_solar_data(sqm=0.4 * 0.5))
     )
 
     # Initialize carbon intensity API
     carbon_api_sim = world.start(
-        "CarbonApi", sim_start=SIM_START, carbon_api=CarbonApi(actual=load_carbon_data())
+        "CarbonApi", sim_start=SIM_START, carbon_api=TimeSeriesApi(
+            actual=load_carbon_data())
     )
     carbon_api_de = carbon_api_sim.CarbonApi(zone="DE")
