+++ tests/test_core.py
@@ -2,32 +2,32 @@ import pandas as pd
 import pytest
 from datetime import timedelta
 
-from vessim.core.simulator import CarbonApi, Generator, TraceSimulator
+from vessim import TimeSeriesApi
 
 
 class TestTraceSimulator:
     @pytest.fixture
-    def trace_sim(self) -> TraceSimulator:
+    def trace_sim(self) -> TimeSeriesApi:
         index = [
             pd.to_datetime("2023-01-01T00:00:00"),
             pd.to_datetime("2023-01-01T00:30:00"),
             pd.to_datetime("2023-01-01T01:00:00"),
         ]
         actual = pd.DataFrame({"a": [1, 2, 3], "b": [0, 3, 0]}, index=index)
-        return TraceSimulator(actual)
+        return TimeSeriesApi(actual)
 
     @pytest.fixture
-    def trace_sim_series(self) -> TraceSimulator:
+    def trace_sim_series(self) -> TimeSeriesApi:
         index = [
             pd.to_datetime("2023-01-01T00:00:00"),
             pd.to_datetime("2023-01-01T00:30:00"),
             pd.to_datetime("2023-01-01T01:00:00"),
         ]
         actual = pd.Series([1, 2, 3], index=index)
-        return TraceSimulator(actual)
+        return TimeSeriesApi(actual)
 
     @pytest.fixture
-    def trace_sim_forecast(self) -> TraceSimulator:
+    def trace_sim_forecast(self) -> TimeSeriesApi:
         index = [
             pd.to_datetime("2023-01-01T00:00:00"),
             pd.to_datetime("2023-01-01T01:00:00"),
@@ -46,7 +46,7 @@ class TestTraceSimulator:
         forecast["request_time"] = pd.to_datetime(forecast["request_time"])
         forecast["forecast_time"] = pd.to_datetime(forecast["forecast_time"])
         forecast.set_index(["request_time", "forecast_time"], inplace=True)
-        return TraceSimulator(actual, forecast)
+        return TimeSeriesApi(actual, forecast)
 
     @pytest.mark.parametrize(
         "dt, expected",
@@ -76,7 +76,7 @@ class TestTraceSimulator:
         assert trace_sim.zones() == ["a", "b"]
 
     def test_trace_actual_at_single_zone(self, trace_sim_series):
-        assert trace_sim_series.actual_at("2023-01-01T00:00:00") == 1
+        assert trace_sim_series.actual("2023-01-01T00:00:00") == 1
 
     @pytest.mark.parametrize(
         "dt, zone, expected",
@@ -90,22 +90,22 @@ class TestTraceSimulator:
         ],
     )
     def test_actual_at(self, trace_sim, dt, zone, expected):
-        assert trace_sim.actual_at(dt, zone) == expected
+        assert trace_sim.actual(dt, zone) == expected
 
     def test_actual_at_fails_if_zone_not_specified(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.actual_at(pd.to_datetime("2023-01-01T00:00:00"))
+            trace_sim.actual(pd.to_datetime("2023-01-01T00:00:00"))
 
     def test_actual_at_fails_if_zone_does_not_exist(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.actual_at(pd.to_datetime("2023-01-01T00:00:00"), "c")
+            trace_sim.actual(pd.to_datetime("2023-01-01T00:00:00"), "c")
 
     def test_actual_at_fails_if_now_too_early(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.actual_at(pd.to_datetime("2022-12-30T23:59:59"), "a")
+            trace_sim.actual(pd.to_datetime("2022-12-30T23:59:59"), "a")
 
     def test_forecast_at_single_zone(self, trace_sim_series):
-        assert trace_sim_series.forecast_at(
+        assert trace_sim_series.forecast(
             start_time=pd.to_datetime("2023-01-01T00:00:00"),
             end_time=pd.to_datetime("2023-01-01T01:00:00"),
         ).equals(
@@ -160,7 +160,7 @@ class TestTraceSimulator:
         ],
     )
     def test_forecast_at(self, trace_sim_forecast, start, end, zone, expected):
-        assert trace_sim_forecast.forecast_at(
+        assert trace_sim_forecast.forecast(
             pd.to_datetime(start), pd.to_datetime(end), zone=zone
         ).equals(expected)
 
@@ -241,7 +241,7 @@ class TestTraceSimulator:
     def test_forecast_at_with_frequency(
         self, trace_sim_forecast, start, end, zone, frequency, method, expected
     ):
-        assert trace_sim_forecast.forecast_at(
+        assert trace_sim_forecast.forecast(
             pd.to_datetime(start),
             pd.to_datetime(end),
             zone=zone,
@@ -251,14 +251,14 @@ class TestTraceSimulator:
 
     def test_forecast_at_fails_if_zone_not_specified(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.forecast_at(
+            trace_sim.forecast(
                 pd.to_datetime("2023-01-01T00:00:00"),
                 pd.to_datetime("2023-01-01T01:00:00"),
             )
 
     def test_forecast_at_fails_if_zone_does_not_exist(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.forecast_at(
+            trace_sim.forecast(
                 pd.to_datetime("2023-01-01T00:00:00"),
                 pd.to_datetime("2023-01-01T01:00:00"),
                 zone="c",
@@ -266,7 +266,7 @@ class TestTraceSimulator:
 
     def test_forecast_at_fails_if_start_too_early(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.forecast_at(
+            trace_sim.forecast(
                 pd.to_datetime("2022-12-31T23:59:59"),
                 pd.to_datetime("2023-01-01T01:00:00"),
                 zone="a",
@@ -274,8 +274,8 @@ class TestTraceSimulator:
 
     def test_forecast_at_fails_with_invalid_frequency(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.forecast_at(
-                trace_sim.forecast_at(
+            trace_sim.forecast(
+                trace_sim.forecast(
                 pd.to_datetime("2023-01-01T00:00:00"),
                 pd.to_datetime("2023-01-01T01:00:00"),
                 zone="a",
@@ -285,47 +285,9 @@ class TestTraceSimulator:
 
     def test_forecast_at_fails_if_not_enough_data_for_frequency(self, trace_sim):
         with pytest.raises(ValueError):
-            trace_sim.forecast_at(
+            trace_sim.forecast(
                 pd.to_datetime("2023-01-01T00:00:00"),
                 pd.to_datetime("2023-01-01T01:00:00"),
                 zone="a",
                 frequency="15T",
             )
-
-
-class TestCarbonApi:
-    def test_initialize_fails_if_unit_unsupported(self):
-        with pytest.raises(ValueError):
-            CarbonApi(pd.DataFrame(), unit="unsupported_unit")
-
-    def test_carbon_intensity_at_with_lb_per_mwh(self):
-        data = pd.DataFrame({"a": [1, 2]})
-        ci_api = CarbonApi(data, unit="lb_per_MWh")
-        assert 0.45 < ci_api.carbon_intensity_at(0) < 0.46
-        assert 0.9 < ci_api.carbon_intensity_at(1) < 0.91
-
-
-class TestGenerator:
-    @pytest.fixture
-    def generator(self) -> Generator:
-        index = [
-            pd.to_datetime("2023-01-01T00:00:00"),
-            pd.to_datetime("2023-01-01T00:30:00"),
-            pd.to_datetime("2023-01-01T01:00:00"),
-        ]
-        data = pd.Series([1, 2, 3], index=index)
-        return Generator(data)
-
-    @pytest.mark.parametrize(
-        "dt, expected",
-        [
-            (pd.to_datetime("2023-01-01T00:00:00"), 1),
-            (pd.to_datetime("2023-01-01T00:00:10"), 1),
-            (pd.to_datetime("2023-01-01T01:00:00"), 3),
-            (pd.to_datetime("2023-01-01T10:00:00"), 3),
-            (pd.to_datetime("2023-01-01T00:29:59"), 1),
-            (pd.to_datetime("2023-01-01T00:30:00"), 2),
-        ],
-    )
-    def test_power_at(self, generator, dt, expected):
-        assert generator.power_at(dt) == expected
