+++ vessim/core/__init__.py
@@ -1,5 +1,1 @@
-"""Vessim core components.
-
-These components, especially the ones in `vessim.core.simulator`, can be used for simple
-experiments that do not require the use of a co-simulation engine.
-"""
+"""Vessim core components."""
